exception handling inprogress

+++ src/main/java/com/example/demo/serviceImpl/StudentServiceImpl.java
@@ -1,6 +1,7 @@
 package com.example.demo.serviceImpl;
 
 import java.util.List;
+import java.util.Optional;
 
 import org.modelmapper.ModelMapper;
 import org.slf4j.Logger;
@@ -11,6 +12,8 @@ import org.springframework.stereotype.Service;
 import com.example.demo.ApiResponse.ResponseMV;
 import com.example.demo.MV.StudentMV;
 import com.example.demo.VM.StudentVM;
+import com.example.demo.exception.CustomExceptionCode;
+import com.example.demo.exception.StudentNotExist;
 import com.example.demo.model.Student;
 import com.example.demo.repository.StudentRepository;
 import com.example.demo.service.StudentService;
@@ -31,6 +34,10 @@ public class StudentServiceImpl implements StudentService {
 	public StudentMV addStudent(StudentVM studentVm) {
 		try {
 			log.info("enter in add student method with request data : {}", studentVm);
+			Optional<Student> findStudent = studentRepository.findById(studentVm.getStudentId());
+			if(!findStudent.isEmpty()) {
+				
+			}
 			Student student = modelMapper.map(studentVm, Student.class);
 			studentRepository.save(student);
 			return modelMapper.map(student, StudentMV.class);
@@ -55,12 +62,18 @@ public class StudentServiceImpl implements StudentService {
 	@Override
 	public StudentMV getStudent(int studentId) {
 		try {
-			log.info("enter in getStudent method for id : {}",studentId);
-			Student student = studentRepository.findById(studentId).get();
+			log.info("enter in getStudent method for id : {}", studentId);
+			Optional<Student> student = studentRepository.findById(studentId);
+			if (student.isEmpty()) {
+				log.error("Student Not Exist : {}", studentId);
+				throw new StudentNotExist(CustomExceptionCode.STUDENT_NOT_FOUND.getErrMsg(),
+						CustomExceptionCode.STUDENT_NOT_FOUND.getErrCode());
+			}
 			return modelMapper.map(student, StudentMV.class);
-		}
-		catch(Exception e) {
-			log.error("exception occurred while find student by id : {}",studentId);
+		} catch (StudentNotExist e) {
+			throw e;
+		} catch (Exception e) {
+			log.error("exception occurred while find student by id : {}", studentId);
 			return null;
 		}
 	}
@@ -68,11 +81,10 @@ public class StudentServiceImpl implements StudentService {
 	@Override
 	public StudentMV getStudentByName(String studentName) {
 		try {
-			log.info("enter in getStudentByName method for name : {}",studentName);
-			Student student = studentRepository.findBystudentName(studentName).get();
-			return modelMapper.map(student, StudentMV.class);
-		}
-		catch(Exception e) {
+			log.info("enter in getStudentByName method for name : {}", studentName);
+			Optional<Student> student = studentRepository.findBystudentName(studentName);
+			return modelMapper.map(student.get(), StudentMV.class);
+		} catch (Exception e) {
 			log.error("exception occurred in getStudentByName method : {}", e);
 			return null;
 		}
@@ -86,8 +98,7 @@ public class StudentServiceImpl implements StudentService {
 			responseMv.setMessage("Record Deleted");
 			responseMv.setStatus(true);
 			return responseMv;
-		}
-		catch(Exception e) {
+		} catch (Exception e) {
 			log.error("exception occurred while deleteStudent for Id : {}", studentId);
 			responseMv.setMessage("Record Deletion Failed");
 			responseMv.setStatus(false);
@@ -98,13 +109,20 @@ public class StudentServiceImpl implements StudentService {
 	@Override
 	public ResponseMV deleteStudentByName(String studentName) {
 		try {
-			log.info("enter in deleteStudent method for id : {}", studentName);
+			log.info("enter in deleteStudentByName method for name : {}", studentName);
+			Optional<Student> student = studentRepository.findBystudentName(studentName);
+			if (student.isEmpty()) {
+				log.error("Student Not Exist with Student Name : {}", studentName);
+				throw new StudentNotExist(CustomExceptionCode.STUDENT_NOT_FOUND.getErrMsg(),
+						CustomExceptionCode.STUDENT_NOT_FOUND.getErrCode());
+			}
 			studentRepository.deleteBystudentName(studentName);
 			responseMv.setMessage("Record Deleted");
 			responseMv.setStatus(true);
 			return responseMv;
-		}
-		catch(Exception e) {
+		} catch (StudentNotExist e) {
+			throw e;
+		} catch (Exception e) {
 			log.error("exception occurred while deleteStudent for name : {}", studentName);
 			responseMv.setMessage("Record Deletion Failed");
 			responseMv.setStatus(false);
